u16 -> i32 q/r/z AzialCoords refact.

+++ client/src/game/systems/input.rs
@@ -100,9 +100,9 @@ pub fn select_entity(
                 q = q.round();
                 r = r.round();
 
-                if q >= 0.0 && r >= 0.0 && q <= u16::MAX as f32 && r <= u16::MAX as f32 {
-                    let q = q as u16;
-                    let r = r as u16;
+                if q >= 0.0 && r >= 0.0 && q <= i32::MAX as f32 && r <= i32::MAX as f32 {
+                    let q = q as i32;
+                    let r = r as i32;
 
                     let qr = AxialCoordinates::new(q, r);
